Fix Pebble push() handling of binary files (#574)
Python's mime code's msg.as_bytes() ends up here:
https://github.com/python/cpython/blob/0f55d21212a0fb6dec4479eb76249ab2b54e57a3/Lib/email/generator.py#L149-L163

But that translates \r\n into \n, which is obviously horrible for
binary files. I saw someone on StackOverflow that had subclassed
Generator and overriden this private _write_lines method, but that
seems like a total hack. Besides, it's not hard to just do this
ourselves and get control over it, so let's do that.

At some point we'll probably add support for streaming, anyway, at
which point all of this will change.