Check whether Pebble is ready before  (#567)
* Add a .ready property to Model.Container to see if Pebble is usable

Try to get a list of services from `Container`. If we get a
`ConnectionError`, return False. This is a first-pass towards a
better workflow for catching common errors when writing charms
and Pebble is not ready yet (or is not ready again)

Closes #553

* Check for ready in more places, add `Container.restart()`

Since there are almost no operations we can actually perform if
Pebble returns a `ConnectionError`, check for it first and log
it if that's the case to give a hint to Charm authors in case they
forget to check.

This required some extra test mocking due to the way
`MockPebbleClient` is implemented.

Add a `restart` method to `ops.model.Container` so authors do not
need to write their own methods for it

* Raise a common exception instead of logging errors if Pebble is not ready, same for getting services

* Use a contextmanager for is_ready() instead

The contextmanager is pretty trival. If it's in the context, log
the actual exceptions thrown up (by type and arguments, but not
the whole backtrace). If not, behave like we currently are

* Add an example for is_ready

* Move is_ready to a separate class instead

* Make _completed a private attribute with a property to fetch it