Fix _should_use_controller_storage for metadata v2 charms.
This PR comprises a hack to fix _should_use_controller_storage for
metadata v2 charms.

A more complete fix relies on not yet implemented facilities in Juju
for determining the underlying cloud a charm is running on.

For now, we mirror a similar hack in Juju core, guessing that we are
on kubernetes if (and only if) this is a metadata v2 charm with a list
of containers.

We place the hack directly inside the CharmMeta class, so that we can
parse the metadata.yaml in a backwards compatible way, and don't have
to completely implement metadata v2 support.