Redefine ops.relation so multi relationships work both ways
Add .ready() and .unready() to ConsumerBase.

multi= is left as an argument to `__init__` for now so we don't
break libraries as soon as it's instantiated, but it does not
actually do anything anymore. Instead, it can now multiplex
in both directions.

Add `.relation` to `ConsumerEvents` by destructuring the relation
and fetching it back once a snapshot is loaded, so `ConsumerEvents`
have `event.relation` like `RelationiEvents` do.

Simplify `ConsumerEvents` so they all inherit from a single base
class, considering that none of them had invidualized behavior.

Remove `TooManyProviders` and add a test to ensure that multiple
consumers can work with one provider.

Closes #586