First pass at adding tox.ini
This is "borrowed" from the observability team. We think that their
tox file serves as an excellent "default" template for the charmed
operator ecosystem.

Running "tox" on this branch will generate a lot of failures at the
moment, however, as I still haven't finished thinking about and
parsing the tox file, and making changes to our codebase.