Implemented Pebble filesystem APIs for ops.testing
ops.testing._TestingPebbleClient has been updated with API functions
mimicking the live Pebble client which interface with a mock backend
filesystem.

The behavior should be more-or-less in line with the live Pebble
client, except in places where it seemed prudent to diverge.  The main
place where we diverge is in regards to permissions and ownership,
where we generally use None as default values for all permissions and
ownership values unless something is explicitly specified by the user.

The mock filesystem combines temporary files on disk with ownership
and permission metadata being stored in memory.

Closes #518