Added error messages to PathErrors in test harness
These error messages are presently not intended to directly match the
messages which would be raised by Pebble, but rather to give a hint as
to what may have gone wrong if these errors were encountered
unexpectedly.  We can of course revisit this if we feel we should have
these error messages line up as closely as possible to what Pebble
generates; I'm not sure how we all feel regarding how "in stone" or
"part of the API" those messages are at this point.