Add storage details to StorageEvents
When any storage event fires, look up the appropriate Storage
object and attach it to the event. Since `juju storage-get` on
any attach or detach will return the type and location without
any additional arguments, it's necessary to iterate through the
list of known storages until we find the location for the appropriate
object.

Once we find it, we can operate by Storage.name and Storage.id for
restoring/deserializing the snapshot data, and further lookups are
not necessary.