[Draft] Initial secrets support.
This is a draft PR. Do not merge!

The final PR will implement Juju secrets support.

This first push largely comprises a copy-paste of the proposed code in
the spec. No unit tests were broken, and the initial round of linter
errors has been fixed :-)

The next step is to add wrappers around the Juju hook commands to
ModelBackend, then wire those in to the public facing methods in this
draft.