Added stubs for actual invocations of the Juju secrets hooks.
I'm not sure that all of these need to be explictly wrapped in the
Operator Framework. And I'd like to move a lot of the fussy arg
parsing and formatting into a Secret class.

But at least I've got things dropped into code, and documented so that
I don't have to switch between google docs and emacs. :-)