Add a method to set planned unit count in the testing harness.
Closes gh#636

In many cases, someone testing code that calls planned units will not
want to make calls to the harness to setup virtual peer units. This PR
provides a nice, relatively clean way to set the return from planned
units (and only the return from planned units) in the harness backend.