Handle bad I/O commands consistently; log warnings on issues
This makes the handling of the I/O command payloads in
_WebsocketReader consistent with _websocket_to_writer, namely
continuing on invalid JSON or an invalid command. It also adds warning
logs when either of these things happen. Both of these behaviors are
what the Go client does, so it's good for consistency too.

Add tests to check these cases are ignored (and logged).