Memory-efficient push via pebble (#666)
* Implemented Pebble filesystem APIs for ops.testing

ops.testing._TestingPebbleClient has been updated with API functions
mimicking the live Pebble client which interface with a mock backend
filesystem.

The behavior should be more-or-less in line with the live Pebble
client, except in places where it seemed prudent to diverge.  The main
place where we diverge is in regards to permissions and ownership,
where we generally use None as default values for all permissions and
ownership values unless something is explicitly specified by the user.

The mock filesystem combines temporary files on disk with ownership
and permission metadata being stored in memory.

Closes #518

* Added error messages to PathErrors in test harness

These error messages are presently not intended to directly match the
messages which would be raised by Pebble, but rather to give a hint as
to what may have gone wrong if these errors were encountered
unexpectedly.  We can of course revisit this if we feel we should have
these error messages line up as closely as possible to what Pebble
generates; I'm not sure how we all feel regarding how "in stone" or
"part of the API" those messages are at this point.

* Improvements based off of code review feedback

* Replaced tempfile-based _File class with a purely in-memory version
  for additional simplicity.
* Replaced explicit permission/ownership arguments with **kwargs when
  appropriate to do so.
* Corrected and tweaked some typing annotations.
* Reformatted some function calls.  Some calls looked large because of
  having one argument per line, but some args (e.g. ownership) had
  logical groups which arguably makes things easier to read if they are
  grouped together.

* Removed unneeded "os" import

* Moved a permissions check for clarity

* More changes per review, plus cleanups

* Added tests for pushing to relative paths; Pebble gives
  generic-file-errors in such cases.
* Reformatted a couple of tests for readability.
* Removed comments I put in place while stubbing out the test cases I
  wanted to handle; they're now redundant.

* Added real pebble backend testing, fixed a few other bits

Methodology borrowed from benhoyt's pending MR; credit goes to him for
the approach.  If RUN_REAL_PEBBLE_TESTS and PEBBLE are appropriately
defined, the pebble-based tests will run.

At the time of writing, one of these tests fails with what might be a
legitimate Pebble-side bug.

* Re-enabled cleanup

* Auto-reformatted

* Removed variable annotation

* Removed trailing comma after kwargs for Py3.5

* Removed other variable annotations

* Removed some unnecesary usages of pathlib.Path in tests

Reasons:
* Pre-Py3.7, pathlib.Paths lack startwith methods, which was allowing
  _MockFIlesystem.create_file to work even if a path object were passed
  in instead of a string for the "path" argument.
* The tests read simpler without them.

* Added Windows support, fixed a name in test teardown

* Revert "Added Windows support, fixed a name in test teardown"

This reverts commit 0150342569e4d699038b5f08614e002cec5f4bdd.

* Fixed TestPebbleStorageAPIsUsingRealPebble.tearDown()

* _MockFilesystem now explicitly uses POSIX-style conventions

Replaced pathlib.Path usage with pathlib.PosixPurePath.

Once Pebble supports Windows, we will want to revisit this based upon the
nuances of Pebble on Windows.

* Replaced PosixPurePath with PurePosixPath

* Fixed docstring per lint

* Changed path types in test checks to be more explicit

* PosixPath -> PurePosixPath

* Lint fixes

* Various cleanups

* Made exception messages start with lowercase
* Removed a stale comment

* Replaced ValueErrors with NonAbsolutePathErrors

* Several review fixes

* Updated FileInfo instantiation to use explicit keyword args.
* Fixed spelling errors. (existant -> existent)

* Skipping a test pending a bug fix

* More review tweaks, and tweaked a test skip

* Replaced pebble.Error "args[0]" references with "kind".
* Handled a test skip in a different way to avoid affecting testing
  against the mock-based backend.

* Replaced __getitem__ and __delitem__

_MockFilesystem isn't really a mapping, but rather a hierarchy, thus it
could be argued that the mapping-like semantics don't quite fit.
Renamed __getitem__ to get_path and __delitem__ to delete_path.

* Fixes for formatting and lint

* Adding test for exercising larger file push/pull

* pebble.Client.push() now pushes large files in chunks

* Centralized chunk size definition

* Some improvements based on review feedback

* Use os.urandom() for creating random data in a simpler way.
* Remove a no-longer-needed nonlocal statement.

* Fixed failing tests in test_pebble.py

This change means that body is now a generator, which means that
_parse_write_multipart has to iterate through the body, rather than
sending it all over to the email parser at once.

Co-authored-by: Jon Seager <jon.seager@canonical.com>
Co-authored-by: Pen Vander Giessen <petevg@gmail.com>