Make the storage harness work when metadata keys have hyphens
Don't replace '-' with '_' until after we find the index. This is
essentially an off-by-one error which prevents any storage key
which uses hyphens from being looked up, and causes a difference
in behavior between actual charm behavior and harness behavior.

Add a small helper to the tests and an additional test to verify
that it works.