remove the temp file from disk before returning it to the caller
The data will remain available until the file is closed (by the user or
the gc).  This way we don't introduce any new API (e.g. a file on disk
the user can do stuff with) - and limit them to reading from returned
object directly.  We can continue to ruminate on temp file locations,
copying, etc.  details until we are confident if/how we want to change
things.