make sure tempfile is removed on windows
On posix platforms, we can unlink the file while it is opened and the os
will clean up the file for us automatically when it is closed either by
the user or by the python gc.  On windows, we can't remove an opened
file.  So instead we use tempfile's internal wrapper class that removes
the file when it is closed/gc'd.  This isn't quite as good as unlinking
the file before returning it because users can still check the returned
file.name attrib and then directly interact with the (unremoved) file on
disk, but it's the best I could come up with for now.