Parse revision|version into CharmMeta
Operator Framework does not currently do anything with `version`
(which no longer appears to be automatically generated by
`charmcraft` anyway) or `revision`, which is always present when
Juju deploys, even if it's `0` for a locally-built charm.

Expose these in `CharmMeta` so authors can use it to
`set_workload_version` based on `charm.meta` information if desired,
or otherwise perform labeling/checking on metrics labels or other
data.