Avoid issuing relation_changed events with no data delta
The test harness will always issue a relation_changed event when
the update_relation_data is invoked with valid values. Juju, instead
will issue relation_changed events to units on the other side of a
relation only when there is an actual "diff" in terms of the content
of the relation data bags.

This commit rectifies the issue :-)

Add test for duplicated relation data update