begin_with_initial_hooks now sets post install status correctly
Change the behavior of the begin_with_initial_hooks test fixture to
better reflect the behavior of production Juju.

Specifically, after the fixture runs the install hook, we check unit
status. If the status is "maintenance" with an empty message, we
assume that the install hook did not set a status, and set the unit's
status to Unknown.

Resolves gh#726