Remove overzealous Harness.add_storage error check. (#781)
When juju runs, storage is attached/mounted into containers before the
charm class instance is constructed - meaning that in production, charms
can access storage data inside their `__init__` functions.  In order to
allow the harness to simulate this circumstance, it must be possible to
attach storage *before* calling Harness.begin().  With that need in
mind, this error check in add_storage doesn't make sense and so is
removed here.

Fixes #780