Added a very basic smoke test. Run it with `tox -e smoke` (#800)
This requires a development box with an active juju controller, and
charmcraft installed.

The smoke test will build the "smoke" charm in test/charms/test_smoke
and try to deploy xenial, bionic and focal versions of the charm.

There are a few wonky things about this, including how we inject the
current ops package, and the fact that we test with just the focal
version of the charm. Future versions of this smoke test should clear
those wonky bits up, as well as add automation.