sketch out what harness secrets impl might look like
The idea is to implement the secret hook tools in the testing backend as
normal, but to have this pseudo-secret object that represents the juju
secret semantics - e.g. tracking which units are on which revisions,
whether a secret is expired, etc.  This allows a test writer to do
things like force-expire a secret at a desired point in time and then
see how a charm reacts to it in various hook calls.  Or to control the
contents or access granting of a secret that is being consumed, and not
explicitly controlled by the charm being tested.