downgrade breaking-change exception to just a warning (#822)
* downgrade breaking-change exception to warning

This harness error check was added in #795.  It raises on a usually
benign sloppy naming of units by charm developers when writing tests.
In order to avoid breaking CI/tests for other charms, this should be
downgraded to a warning only. Notably, this change does/did cause
breakage in some data platform and observability charms - so it isn't
just a theoretical problem.

* see if we can fix typing errors

* fix over-zealous relation_get error on harness

* typing fix attempt