Clarify what Container.get_plan actually returns (#864)
This doc tweaks is to clarify for #863.

Note that the Sphinx change is because 3.x no longer installs on recent
Python versions. I updated this to fix doc build on Python 3.10.x. Due
to changes in the typing module in Python 3.10.x and how Sphinx was
doing a version test using version > (3, 10), doc building was broken
with an ImportError when import Sphinx on Python 3.10.x (I'm using
Python 3.10.6, for example). See: sphinx-doc/sphinx@8b2031c ... So just
use the latest version of Sphinx for now.

Fixes #863