Remove SIMULATE_CAN_CONNECT, make begin_with_initial_hooks trigger pebble-ready (#873)
When we introduced `can_connect` it would have been a breaking change
to enable it by default, so we added the `SIMULATE_CAN_CONNECT`
global to enable it. The default was False to avoid breaking changes.
We've been planning on turning this on by default for the 2.0
release (basically the only breaking change for 2.0). Note that it's
been a recommendation and a warning to date.

This PR removes the global flag and enables simulation by default. The
default value of `container.can_connect()` is still False when using
`begin()` ("manual mode"), but this PR updates
`begin_with_initial_hooks()` to set `can_connect` to True and fire
pebble-ready for all containers in the charm's metadata.