Make testing pull() raise same error as real system (pebble.PathError)
Currently the testing harness's version of pull() raises
FileNotFoundError and IsADirectoryError instead of pebble.PathError,
which is what the real pull() raises in these situations. Catch those
and raise the same error as in real life.

Also document this better in the relevant docstrings.

There are probably other places the testing file system's errors leak
out, but keep this PR limited in scope to just pull().

Also remove an unnecessary Pyright FIXME -- this was addressed when
the message() method was removed recently.

Fixes #896