Revert to static list for setup.py install_requires
This breaks "pip install" because requirements.txt is not included in
the sdist or wheel. We could explicitly add a MANIFEST.in, but seems
better to keep it simple and in line with the defaults, so just
duplicate it and add a test to ensure they stay in sync.

Fixes #913.