Explicitly add all public ops.X names to __all__ in __init__.py
This is so that when using "import ops; ops.X" in a charm, Pyright
and MyPy won't complain (in the charm's codebase).

For example, MyPy:

src/charm.py:14: error: Name "ops.CharmBase" is not defined  [name-defined]

And Pyright:

src/charm.py:32:50 - error: "CharmBase" is not exported from module "ops" (reportPrivateImportUsage)

This also avoids disabling the reportUnusedImport setting.

I was considering adding a test that checked that __all__ stays in sync
with the imports below, but Pyright errors if we have a entry in one
but not the other.