Lock docs dependencies; build docs with recent Sphinx version (#938)
I noticed that my local build of the docs was significantly different
from the ReadTheDocs one, so I wanted to get them in sync.

Per https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html,
this updates the .readthedocs.yaml config for the latest format, pins
the dependency versions (transitively).

To do the transitive pinning, use a requirements.in (which should be
updated by hand) that generates a fully-pinned requirements.txt (which
shouldn't be edited). Added the docs/update_requirements.sh script to
do this.

Now that we're on a recent Sphinx version, we need to use
autodoc_typehints='signature' to avoid the extra empty "Parameters"
sections on all classes.