Use Canonical-ish theme from canonical/sphinx-docs-starter-pack (#941)
It's not perfect, but I think it's probably an improvement to be more
consistent with Canonical styling. We'll update to the
Canonical "Vanilla" theme once the web & design team are ready with
that.

I definitely don't love the `_compute_navigation_tree` patch we have
to do in `conf.py` (it seems like there's an oversight in the Furo
theme to not fall back to the "local TOC" like the sphinx-rtd-theme
did). But it will do the job for now, and hopefully the Canonical
theme will address it properly.

You can view the rendered version here:
https://ops--941.org.readthedocs.build/en/941/