Remove jargon from and simply README
Remove a lot of the "operator" and K8s jargon from the top of our
README. I feel fairly strongly this isn't the place to sell how great
Juju, K8s operators, and model-driven development is -- we should do
that on our landing page and/or docs, not in the ops README.

Also simplify and update some of the terminology (hopefully consistent
with our updated terminology doc, though I've used the "charmed" prefix
as there's more than one "Operator Framework").

The first part of the doc is copied to ops/__init__.py for consistency.
That appears at the top of our API reference at https://ops.readthedocs.io/

Move the note about depencies to HACKING.md where I think it belongs.