Various docstring and type annotation updates (#953)
The intent here is to sprinkle a bit of consistency over our reference
docs. There was previously a fair bit of duplication, lack of type
information on attributes, or just unclear docs, and so on. One of
the biggest changes is giving instance attributes proper type
annotations and separating out and enhancing their docstrings.

It's not perfect, but I think it's a definite improvement for #920.