Add Secret.unique_identifier (just the XID part)
We want to provide a way to allow charms to test two secrets for
equality / see if they refer to the same secret object.

It would be nice if Secret.id was a unique identifier, but it's really
a locator URI. Too late to change that, though.

We considered deprecating Secret.id entirely, and adding two new
properties, "uri" and "unique_identifier". However, given that this
locator is called ID throughout the rest of the codebase and by Juju,
I think that would introduce more confusion.

The name "unique_identifier" is long on purpose: it should be
relatively rare to use it, and we want to nudge people to using the
locator (.id) and labels instead.

See context at https://bugs.launchpad.net/juju/+bug/2028402