Add Secret.unique_identifier (XID part); improve id/label docs (#986)
In some cases when a charm has an array or set of secrets that don't
have distinguishing characters, charms need a way to distinguish by
ID (most of the time you'd use labels: most secrets have a specific
use like "database-password" that you can use as a label). Despite
the name, you can't `Secret.id` as a unique identifier, because
sometimes it includes the model UUID, and sometimes not -- it's
really intended as a locator URI, not a unique ID. So provide
`Secret.unique_identifier` to handle this use case.

It would be nice if Secret.id was a unique identifier, but it's really
a locator URI. Too late to change that, though.

We considered deprecating Secret.id entirely, and adding two new
properties, "uri" and "unique_identifier". However, given that this
locator is called ID throughout the rest of the codebase and by Juju,
I think that would introduce more confusion.

The name "unique_identifier" is long on purpose: it should be
relatively rare to use it, and we want to nudge people to using the
locator (.id) and labels instead.

See more context at https://bugs.launchpad.net/juju/+bug/2028402