Ignore push() encoding if source is a binary file or stream
The push() docs explicitly say that the "encoding" param is ignored if
the source is bytes or a binary stream. This is useful because the
default is encoding='utf-8'; but we should already know the source
stream is binary, so let that override the default.

Also add a test that "encoding" is ignored in these cases.

Fixes #990