Remove the PR template
It just gets in the way. You have to manually select and delete it
from the PR description -- and that's easy to miss when you have a
single commit and GitHub puts this text at the bottom of the nice
commit message.

More trouble than it's worth -- just delete it. You can be sure I'll
ask the PR creator if it needs a better description. :-)