Remove the type hitn overloads.
In other repos, there's code like this:

Unfortunately, pyright cannot tell that the protocol argument is always 'tcp' in this case, only that the protocol is one of tcp/udp/icmp, and that's too broad to fit the righter overloaded specs, so it complains. This could be handled with changes in the upstream places, but I don't see any simple way to handle it within ops itself.

This was a minor improvement and not really needed, plus it's in the two methods that we're wanting to move people away from, so drop this change.