Fix type hints.
For the model backend stub, we know that the type won't match, but we don't really care. We could make an abc that both the stub and the real class could inherit from, or we could subclass for the stub, but those seem overkill for this specific use case. It seems unlikely that type hints would show any real-life issues here.