test: fix type hint warnings in test_private tests (#1010)
* Silence type warnings.

We are explicitly testing passing the wrong type here, so want this to be incorrect.

* Remove presumably accidental double-call.

Looking more closely, I don't believe the double call was intended. The test still passes because the inner call raises the ValueError all the way up, but that means the outer call is never actually used, and it's a bit weird to be testing that passing a datetime.datetime doesn't work (and it would raise a TypeError anyway).