feat: add Unit.set_ports() for declarative port opening (#1005)
* Make the protocol argument of open_port and close_port optional.

In the majority of cases, the protocol is tcp, so make it a little easier and more concise for those cases.

* Rename OpenPort to Port.

This avoids the impression that the port is already/always opened, rather than just being a port and the instruction is then given via the open/close methods.

* Add a Unit.set_ports method to declare which ports should be open.

Tests still to be added.

This does do a transform to Port() objects that are then disassembled again almost immediately, for the case where ints are provided. We could normalise to a pair of (str, int) instead to avoid that. Not normalising at all ended up with a fairly messy couple of loops instead of simple set operations.

* Verify that the set_ports() method works.

This can be flakey, because we're working with sets but then asserting that things happen in a specific order. That should get resolved.

* Fix flakiness.

The opened-ports call should always be first, but after that it depends on the order of items returned from set operations, and we don't care which order they happen in, so compare the sorted items.

* Normalise to (str, int) to avoid unnecessary transforms.

* Enforce structure on the open_port and close_port methods.

Add typing overloads to ensure that type checking will raise issues.

Good:
 * open_port('icmp')
 * open_port('tcp', 1)
 * open_port('udp', 2)
 * open_port(port=3)

Bad:
 * open_port()
 * open_port('icmp', 1)
 * open_port('tcp')
 * open_port('udp')
 * open_port(4, 'tcp')

Also check that nothing bad has been done at runtime, raising an error (ModelError to be consistent with ops.testing) if one of the invalid specifications is used.

* Remove the default protocol of tcp, as per discussion.

Also some minor adjustments per code review comments.

* Remove overload docstrings and satisfy the linter with #noqa instead.

* Remove the type hitn overloads.

In other repos, there's code like this:

Unfortunately, pyright cannot tell that the protocol argument is always 'tcp' in this case, only that the protocol is one of tcp/udp/icmp, and that's too broad to fit the righter overloaded specs, so it complains. This could be handled with changes in the upstream places, but I don't see any simple way to handle it within ops itself.

This was a minor improvement and not really needed, plus it's in the two methods that we're wanting to move people away from, so drop this change.

* Minor docstring fix.

* Update docstring to align with the new open/close port ones.

* Remove empty file, presumably a bad git add.

* Remove duplicated validation.

* Split tests into separate test methods.

* Tweak docstring: "&" -> "and"

---------

Co-authored-by: Ben Hoyt <ben.hoyt@canonical.com>
Co-authored-by: Ben Hoyt <benhoyt@gmail.com>