test: add type hints to pebble-related tests (#1030)
* Adds overloads for pebble's `pull` method to clarify what you get when you do/do not provide an encoding.
* `test/pebble_cli.py` has a bit of a workaround - the type returned by `exec` is generic, and I believe we could have pyright understand which one was being used, but by repeating the `exec` call in the various `if` statements above. It seems like this would make the code messier for no real gain, so I've cheated and told pyright it will be a `[str]` even though that may not be true.
* Add a few asserts to show pyright that values aren't None or are the correct type
* Add `type: ignore' where we are checking that types are invalid
* Sprinkle type hints as required for pyright to understand what's happening

Note that the latest version of pyright has a problem where we use code like this:

```python
with self.assertRaises(ops.pebble.ExecError) as cm:
    call_something()
self.assertEqual(cm.exception, some_value)
```

The issue is related to `ExecError` being generic combined with `cm` being generic and dynamic. The version of pyright we are currently using does not have an issue with this - but looking at the pyright release notes (1.1.318) it seems like accepting this is considered a bug and failing it in the latest version is correct. However, it's not clear what we are meant to do here, and the documentation (stdlib and pyright) is vague, so I'm leaving this to figure out (if necessary) when we upgrade our pyright version.

We now type check everything under `test/`, except for the smoke test (which needs to wait until pytest-operator and pylibjuju have more type hinting).

Fixes #1007