docs: document more of the exceptions that may be raised (#1044)
### More extensively document the exceptions that Charmers should be aware of and potentially catching

If we raise an exception when checking an argument's type, and the type hinting covers this, I have not included it in the `Raises` list. I feel that the docs (and a type checker) cover this already, and the `Raises` section should be more focused on exceptions that the Charmer should consider handling.

#908 goes back-and-forth on documenting very common exceptions (like `APIError` and `CommunicationError` in nearly all pebble `Client` methods), and I have as well. ~I'm currently +0 on documenting them, and it seems like it's simpler to add them all in the PR and them remove them than decide to add them after the initial PR, so they are included at the moment. I could easily be convinced that they should be removed and a central doc note added (but I'm not sure where that would be best).~ These have been removed and are centralised in the `Container` and `pebble.Client` docstrings.

Exceptions that are possible but should be rare and indicate a broken system (such as Juju or Pebble returning invalid data or being unresponsive) are not included. It feels like this tips *too* far in the "document everything" direction, and I don't feel like it would be worth Charmers trying to handle these exceptional (no pun intended) circumstances.

Where it was unclear whether it was worth documenting a potential exception - such as not providing any commands in `exec`, which seems more likely to be a coding error than something the Charm should handle, but I could see happening with quite dynamic code - I have generally taken the more defensive approach of including the potential raise.

I wrote a terrible, horrible, no good, very bad charm that causes a large number of these errors, and examined the pebble & Juju code to find information about what might cause issues. However, since I'm still only lightly familiar with Juju/ops, I could have missed some cases.

### Minor tweaks to exception chaining

* If we are raising an exception in an exception handler but the original exception is just an implementation detail, `raise x from None`
* If we are converting an exception to a more specific/localised one and the original exception is relevant, `raise x from e`
* If we are filtering an exception and then re-raising, `raise` or `raise e` (preferably the former, but leaving the latter if already in place)

### Miscellanous

One minor type hinting improvement: if an exception is always raised, use `typing.NoReturn` as the return type.

Drive-bys:
 * Remove one `type: ignore` that is no longer required.
 * Add `ops.` to a couple of examples in the docstrings that were missed in #921

Fixes #908, #1043